Solaris.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@4474 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r0drv/solaris/time-r0drv-solaris.c
@@ -27,7 +27,7 @@
 
 RTDECL(uint64_t) RTTimeNanoTS(void)
 {
-    return gethrtime();
+    return (uint64_t)gethrtime();
 }
